Problems creating the assets
To fix the problem with the assets bundle, the extract-text-webpack-plugin had to be updated to the next major version with the @next switch. This fixes the bundle error I had (and I also see on your DigitalOcean installation: Error 404 on app.js / style.css). The command needed is:
```
npm install extract-text-webpack-plugin@next
```

this fixes the problem for those who only want to update instead of a fresh clone or pull.